pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.5"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.8"
djangorestframework-simplejwt = "^5.3.0"
django-filter = "^23.3"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]